ext2 → ext4: 64-bit Timestamps, 256-Byte Inodes. 2254/97.
- Rename ext2 → ext4 durchgehend (Dateien, Funktionen, Defines, Tests)
- ext4_inode_extra: i_ctime_extra/i_mtime_extra/i_atime_extra (Linux ext4 Layout)
- ext4_read_extra/ext4_write_extra fuer 256-Byte Inodes
- ext4_encode_ts/ext4_decode_ts: 34-bit Timestamps (epoch bis 2446)
- vfs_inode: atime/mtime/ctime uint32_t → uint64_t
- fill_ext4_stat: liest Extra-Felder fuer 64-bit stat Timestamps
- utimensat: schreibt Extra-Felder bei 256-Byte Inodes
- mkalpine.sh: mkfs.ext4 -I 256 (256-Byte Inodes aktiviert)
- ENOTSUP, EXDEV, ENODEV in errno.h

+++ src/kernel/fs/tmpfs.c
@@ -1,6 +1,6 @@
 /* CosmoRT tmpfs — in-memory filesystem
  *
- * Mounted at /dev/shm, /tmp, and as fallback root when no ext2.
+ * Mounted at /dev/shm, /tmp, and as fallback root when no ext4.
  * Uses vfs_node/vfs_inode dentry tree from the VFS layer.
  */
 
@@ -226,9 +226,9 @@ static int tmpfs_op_utimensat(struct mount *mnt, const char *relpath,
         uint32_t now = timer_epoch_sec();
         int64_t atime_nsec = times[1], mtime_nsec = times[3];
         if (atime_nsec != UTIME_OMIT)
-            inode->atime = (atime_nsec == UTIME_NOW) ? now : (uint32_t)times[0];
+            inode->atime = (atime_nsec == UTIME_NOW) ? now : (uint64_t)times[0];
         if (mtime_nsec != UTIME_OMIT)
-            inode->mtime = (mtime_nsec == UTIME_NOW) ? now : (uint32_t)times[2];
+            inode->mtime = (mtime_nsec == UTIME_NOW) ? now : (uint64_t)times[2];
     } else {
         uint32_t now = timer_epoch_sec();
         inode->atime = now;
